Less specific roles
Rather than relying on specifically named roles, we're now looking for types of roles (e.g. teacher and student), which allows us to -- for example, rename the teachers in our advisory groups as 'advisors' and the students as 'advisees' -- but the dashboard doesn't need to know or care.

Also fixed caching of the observers/observees to make it course-specific (which seems to improve response time).